ord-lsp: move the analysis engine into the lsp package
